Use timeline model to compute availability.
DeviceCounter makes use of intervals stored in SortedSets and SortedDicts to
construct a timeline for computing availability.

Instead, use a single SortedKeyList that measures the deltas of devices being
added/removed and compute the number of devices per interval once. Not only
is it simpler, it has a lower algorithmic complexity due to avoiding constant
item removal/addition when subdividing intervals and thus runs considerably
faster.